remove sst1m event source and use plugin instead
