+++ ctapipe/io/lsteventsource.py
@@ -122,7 +122,7 @@ class LSTEventSource(EventSource):
 
     @staticmethod
     def is_compatible(file_path):
-        from .sst1meventsource import is_fits_in_header
+        from .utils import is_fits_in_header
         if not is_fits_in_header(file_path):
             return False
